requirements.txt
absl-py==1.4.0
astunparse==1.6.3
autopep8==1.6.0
bcrypt==3.2.2
black==22.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.1
Flask-WTF==1.0.1
flatbuffers==23.1.21
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
Sastrawi==1.0.1
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.2.0
urllib3==1.26.14
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
